fix: Don't make "Controllers" when possible
https://hono.dev/docs/guides/best-practices
